server/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-schedule-server"
version = "0.1.0"
description = "Backend skeleton for the smart schedule MVP."
requires-python = ">=3.10"
dependencies = [
  "alembic>=1.13,<2.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "pgvector>=0.3,<1.0",
  "psycopg[binary]>=3.2,<4.0",
  "pytest>=8.0,<9.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "pydantic-settings>=2.0,<3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

